move tests to one file
